Ajsutes de usuarios web y usuarios de tipo movil.

+++ utn_sendero_seguro_web/src/main/java/mx/edu/utn/senderoseguro/web/controlador/dto/SolicitudBase.java
@@ -9,6 +9,6 @@ public class SolicitudBase implements Serializable {
 	
 	private static final long serialVersionUID = 1L;
 	
-	private Long idUsuarioSolicitante;
+	private Integer idUsuarioSolicitante;
 
 }
